+++ services/web/server/tests/unit/with_dbs/03/conftest.py
@@ -5,9 +5,11 @@
 
 
 from collections.abc import AsyncIterator
+from typing import Any
 
 import pytest
-from pytest_simcore.helpers.monkeypatch_envs import setenvs_from_dict
+import sqlalchemy as sa
+from pytest_simcore.helpers.monkeypatch_envs import EnvVarsDict, setenvs_from_dict
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from simcore_postgres_database.models.user_preferences import user_preferences_frontend
 from sqlalchemy.ext.asyncio import AsyncEngine
@@ -36,3 +38,41 @@ def app_environment(
             "WEBSERVER_TRACING": "null",
         },
     )
+
+
+@pytest.fixture
+async def support_group_before_app_starts(
+    asyncpg_engine: AsyncEngine,
+    product_name: str,
+) -> AsyncIterator[dict[str, Any]]:
+    """Creates a standard support group and assigns it to the current product"""
+    from pytest_simcore.helpers.postgres_tools import insert_and_get_row_lifespan
+    from simcore_postgres_database.models.groups import groups
+    from simcore_postgres_database.models.products import products
+
+    # Create support group using direct database insertion
+    group_values = {
+        "name": "Support Group",
+        "description": "Support group for product",
+        "type": "STANDARD",
+    }
+
+    # pylint: disable=contextmanager-generator-missing-cleanup
+    async with insert_and_get_row_lifespan(
+        asyncpg_engine,
+        table=groups,
+        values=group_values,
+        pk_col=groups.c.gid,
+    ) as group_row:
+        group_id = group_row["gid"]
+
+        # Update product to set support_standard_group_id
+        async with asyncpg_engine.begin() as conn:
+            await conn.execute(
+                sa.update(products)
+                .where(products.c.name == product_name)
+                .values(support_standard_group_id=group_id)
+            )
+
+        yield group_row
+        # group will be deleted after test
